backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
earthengine-api==0.1.410
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
